Some branding additions.
Added a few pre-built groups and scales and they are all immutable. Put their data in teh strings.xml file.

You can view the notes on a chart by moving the vertical select bar to a date and clicking the list button.

Rewrote the code to pull result values as a SQL query. (a bit faster). Classes have been abstracted to allow for the creation of a provider interface so charts can pull their data when they are displayed for the first time.





git-svn-id: http://svn.t2.local/svn/mobile/trunk/android/VAS/VAS@312 9a4616e5-df29-b34f-a986-34b38b490834