Features. Long pressing a day in the results chart will bring up the add notes interface.
Bug fixes:
Health tips only show up when the main activity first loads.
Notes indicators are visible on charts despite the existence of data on that day.
Moved the form submission button to the footer of the listview in the forms activity.
installing of test result data no longer adds data for the current date. Now you can actually use the app after installing it for the first time.

git-svn-id: http://svn.t2.local/svn/mobile/trunk/android/VAS/VAS@335 9a4616e5-df29-b34f-a986-34b38b490834